[JSC] IPInt slow path for `memory.atomic.notify` truncates the Memory64 pointer and offset to 32 bits https://bugs.webkit.org/show_bug.cgi?id=314260
Reviewed by Keith Miller.

The IPInt assembly pushes the i64 address operand and the 64-bit immediate
offset to the stack as full 64-bit values, but the C++ slow path reads them
back through the `.i32` member of the IPIntStackEntry union, dropping the
upper 32 bits. For Memory64, an out-of-bounds address such as 2^32 truncates
to 0 and the bounds check is fooled into accepting it.

Test: JSTests/wasm/stress/memory64-atomic-notify-out-of-bounds.js

* JSTests/wasm/stress/memory64-atomic-notify-out-of-bounds.js: Added.
* Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.cpp:
(JSC::IPInt::WASM_IPINT_EXTERN_CPP_DECL):

Canonical link: https://commits.webkit.org/312831@main

+++ Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.cpp
@@ -1284,10 +1284,10 @@ WASM_IPINT_EXTERN_CPP_DECL(memory_atomic_wait64, IPIntStackEntry* args)
 WASM_IPINT_EXTERN_CPP_DECL(memory_atomic_notify, IPIntStackEntry* args)
 {
 #if CPU(ARM64) || CPU(X86_64)
-    unsigned offset = args[0].i32;
+    uint64_t offset = args[0].i64;
     uint8_t memoryIndex = args[1].i32;
     int32_t count = args[2].i32;
-    unsigned base = args[3].i32;
+    uint64_t base = args[3].i64;
     int32_t result = Wasm::memoryAtomicNotify(instance, base, offset, count, memoryIndex);
     WASM_RETURN_TWO(std::bit_cast<void*>(static_cast<intptr_t>(result)), nullptr);
 #else
